backened/requirement.txt
flask
flask-cors
langchain
langchain-google-genai
langgraph
python-dotenv
requests
pydantic